added avg purchases/deposits

+++ finance-backend/src/main/java/com/example/finance_backend/User/StoreData.java
@@ -5,7 +5,7 @@ public class StoreData {
     private String vendor;
     private double income;
     private double expenses;
-    private double numDeposits;
+    private int numDeposits;
     public StoreData(String vendor) {
         this.vendor = vendor;
         this.expenses = 0;
@@ -16,11 +16,11 @@ public class StoreData {
 
     public String getVendor() {return vendor;}
     public double getIncome() {return income;}
-    public Double getExpenses() {return expenses;}
+    public double getExpenses() {return expenses;}
     public int getNumPurchases() {return numPurchases;}
-    public double getNumDeposits() {return numDeposits;}
+    public int getNumDeposits() {return numDeposits;}
 
-    public void setNumDeposits(double numDeposits) {this.numDeposits = numDeposits;}
+    public void setNumDeposits(int numDeposits) {this.numDeposits = numDeposits;}
     public void setExpenses(double expenses) {this.expenses = expenses;}
     public void setIncome(double income) {this.income = income;}
     public void setNumPurchases(int numPurchases) {this.numPurchases = numPurchases;}
